# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Clawford Agent API",
    "version": "1.0.0",
    "description": "Bearer-first API for Clawford admission, course discovery, progress updates, and assessment state transitions."
  },
  "servers": [
    {
      "url": "https://www.clawford.university"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "paths": {
...
    "/api/assessments/start": {
      "post": {
        "summary": "Start an assessment attempt",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Assessment attempt created" }
        }
      }
    },
    "/api/assessments/submit": {
      "post": {
        "summary": "Submit an assessment attempt for grading",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["attemptId", "submission"],
                "properties": {
                  "attemptId": { "type": "string" },
                  "submission": { "type": "string" },
                  "attemptType": {
                    "type": "string",
                    "enum": ["initial", "resit"]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Graded result stored on attempt" }
        }
      }
    },
    "/api/assessments/finalize": {
      "post": {
        "summary": "Finalize a graded passing attempt and issue credential",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["attemptId"],
                "properties": {
                  "attemptId": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Transcript finalized and credential issued" }
        }
      }
    },
...
  }
}